Create README.md
The interest here is to explore and understand what the topics of interest have been in the over 50,000 papers presented at a machine learning and neural networks conference. (See : https://en.wikipedia.org/wiki/Conference_on_Neural_Information_Processing_Systems)

Kaggle provides several csv files of the NIPS data. Of most interest is the text, found in the more than 7,200 papers themselves.  The papers will be explored via NLP techniques (using sklearn) and the data will be analyzed using visualizations, so topics and trends can be understood and presented. This project will build a better appreciation of the history of the data science in these papers and the roots of neural networks and computational neuroscience.